% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{build_data_for_dewarping}
\alias{build_data_for_dewarping}
\title{Create data for Bayesian 2D dewarping}
\usage{
build_data_for_dewarping(score_mat_list, analysis_id, rf_grid, N_g_vec)
}
\arguments{
\item{score_mat_list}{A list of matrices. Each matrxi has B rows and N_g columns. NA means non-peaks, a number,
usually 3, represents peaks.}

\item{analysis_id}{choose the gels in this analysis; user's choice.}

\item{rf_grid}{A vector of length B (=670) equal-spaced grid from 0 to 1.}

\item{N_g_vec}{A named numeric vector showing the number of lanes (incl. Lane 1s)
for each gel with the name being the original ID before being subset by analysis_id.}
}
\value{
A data frame comprised of 5 elements:
\itemize{
\item \code{gel_ID} 
\item \code{lane_ID}
\item \code{band_ID} 
\item \code{Y} Location on Gel for the peaks
\item \code{peak_bin_index} The bin index for the peaks
\item \code{lane_ID_stacked} cumulative counts.
}
}
\description{
Construct data frame that store all detected bands (NB: the result can be trimmed to exclude specific rows.)
}
