% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-peak.R
\name{compute_local_relative_score}
\alias{compute_local_relative_score}
\title{Compute local relative scores:}
\usage{
compute_local_relative_score(dat_binned, half_width, lanes,
  elev_vec = rep(0.014, length(lanes)))
}
\arguments{
\item{dat_binned}{Binned autoradiointensity data}

\item{half_width}{See \code{\link{edges}}}

\item{lanes}{The index of lanes to be analyzed}

\item{elev_vec}{See the argument \code{elevation} in \code{\link{sign_score}}.Default 0.014.}
}
\value{
a matrix of relative scores (nrow=number of bins, ncol=number of lanes analyzed)
}
\description{
Compute using sign_score for each lane to find peak candidate regions
}
