% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-peak.R
\name{compute_score_mat_peak}
\alias{compute_score_mat_peak}
\title{Get the bins for bands:}
\usage{
compute_score_mat_peak(score_mat, score_thres, dat_binned,
  gap_considered_identical_peak = 5)
}
\arguments{
\item{score_mat}{The matrix of scores calculated by \code{\link{compute_local_relative_score}}}

\item{score_thres}{The threshold greater than or equal to which we call the peak-candidate regions (
must be contiguous)}

\item{dat_binned}{Binned autoradiointensity data}

\item{gap_considered_identical_peak}{Default is 5 bins. To merge nearby contiguous peak-candidate
regions if they are less than or equal to \code{gap_considered_indentical_peak} away.}
}
\value{
A matrix with \code{NA} and \code{score_thres}. Those entries equaling \code{score_thres} are bins for peaks.
}
\description{
Given each contiguous peak candidate regions, find the bin that we call peak (aka band)
}
