% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewarp.R
\name{correct_batch_effect_by_ref}
\alias{correct_batch_effect_by_ref}
\title{Piecewise-linear warping to correct batch effects}
\usage{
correct_batch_effect_by_ref(dat_binned_query, dat_binned_ref,
  peak_ind_mat_query, peak_ind_mat_ref, bp, score_thres)
}
\arguments{
\item{dat_binned_query}{The binned data to be compressed or stretched towards the reference}

\item{dat_binned_ref}{The binned data to be aligned to. It is usually arbitrarily chosen
or the one least visually warped (could be very subjective). We suggest in one particular
analysis, fix this data to be one specific gel.}

\item{peak_ind_mat_query}{A matrix of values, with the first-column entries that equal \code{score_thres} considered
as peaks in the query lane}

\item{peak_ind_mat_ref}{Same as \code{peak_ind_mat_query} but for the reference gel
to be aligned to.}

\item{bp}{The break points that binned the smoothed data into bins.}

\item{score_thres}{The score threshold above or equal to which we call peak candidate bins.}
}
\value{
A list of two elements, the first is the vector of break points at reference
peak locations (large to small; NB:check this), the second is a vector of the same length
as the length of reference lane, with each element \code{warped_ind[i]} being the index in the query lane
that corresponds for reference index \code{i}.
}
\description{
Correct the error that the same sample with known molecules can show up at
slightly different locations when autoradiographed on multiple gels.
}
