% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewarp.R
\name{create_batch_aligned_data}
\alias{create_batch_aligned_data}
\title{Create the binned data after piecewise linear warping}
\usage{
create_batch_aligned_data(dat_binned_list, score_mat_peak_list, query_id,
  ref_id, bp, score_thres)
}
\arguments{
\item{dat_binned_list}{A list of data each being the binned data}

\item{score_mat_peak_list}{A list of score matrix with peak locations filled
with numeric values}

\item{query_id}{The gel number that is to be aligned to the reference gel}

\item{ref_id}{The number of the reference gel}

\item{bp}{The break points that binned the smoothed data into bins.}

\item{score_thres}{The score threshold above or equal to which we call peak candidate bins.}
}
\value{
We can name the results \code{dat_binned_batch_aligned_list}
}
\description{
Create the binned data after piecewise linear warping
}

