% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{crop_end}
\alias{crop_end}
\title{remove the right end of the gel image}
\usage{
crop_end(dat_binned, percent = 4.4, bp)
}
\arguments{
\item{dat_binned}{Binned autoradiointensity data}

\item{percent}{Default is 4.4. The percentage of bins at the right end to be cropped.}

\item{bp}{Break points that defined the binning in \code{\link{bin_gel}}.}
}
\value{
A list of two elements:
\itemize{
\item \code{dat_binned_cropped} Binned data after cropping
\item \code{break_points_cropped} Break points after cropping
}
}
\description{
Generally performed before pattern recognition, because light molecules cannot
be distinguished and hence clustered at the righ end in the experiment.
}

