% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewarp.R
\name{dewarp2d}
\alias{dewarp2d}
\title{Fit Bayesian 2D image dewarping models}
\usage{
dewarp2d(data_dewarp, mcmc_options)
}
\arguments{
\item{data_dewarp}{Data frame with columns: gel_ID (true id), lane_ID (1 to say 20, 1 will be removed),
band_ID (counts within each lane), Y (t-scale location of peaks), peak_bin_index (index of the bin corresponding
to the peaks; requires common binning across images), lane_ID_stacked (cummulative lane number upon stacking
gels one-by-one; mainly for plotting).}

\item{mcmc_options}{A list of Markov chain Monte Carlo (MCMC) options.

\itemize{
\item \code{debugstatus} Logical - whether to pause WinBUGS after it finishes
model fitting;
\item \code{n.chains} Number of MCMC chains;
\item \code{n.burnin} Number of burn-in samples;
\item \code{n.thin} To keep every other \code{n.thin} samples after burn-in period;
\item \code{result.folder} Path to folder storing the results;
\item \code{bugsmodel.dir} Path to WinBUGS model files;
}}
}
\value{
BUGS fit results.
}
\description{
Fit Bayesian 2D image dewarping models
}
\details{
This function prepares data, specifies hyperparameters in priors, 
initializes the posterior sampling chain, writes the model file (for JAGS syntax), 
and fits the model. Features:

If running JAGS on windows, please go to control panel to add the directory to
jags into ENVIRONMENTAL VARIABLE!
}
\concept{model fitting functions}
