% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-peak.R
\name{edges}
\alias{edges}
\title{Get the two-sided end points}
\usage{
edges(i, half_width, n)
}
\arguments{
\item{i}{i-th bin}

\item{half_width}{The number of bins to the left or right of the current bin (specified by
parameter \code{i})}

\item{n}{The total number of bins}
}
\value{
A vector of length two, representing the left and right interval end points
}
\description{
Centered at the i-th bin, find the bin half_width away to the left and to the right;
Truncate at the 1st or the biggest bin when hitting the boundary. Essential
for evaluating if a bin is likely to be the location of a peak (aka band)
}
