% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{image_gel}
\alias{image_gel}
\title{Make grayscale heatmap for one gel}
\usage{
image_gel(dat_raw, band_at = NULL, ...)
}
\arguments{
\item{dat_raw}{Raw autoradiointensity data as obtained from \code{\link{read_IPdata}}}

\item{band_at}{A vector of integers selected from 1 to the total number of lanes to highlight selected lanes}

\item{...}{other graphical parameters}
}
\value{
A heatmap visualizing the autoradiointensity data obtained from a single gel.
}
\description{
Make grayscale heatmap for one gel
}
