% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myARI}
\alias{myARI}
\title{Compare two partitions (could be of different lengths)}
\usage{
myARI(x, y)
}
\arguments{
\item{x}{cluster indicators for the first partition}

\item{y}{cluster indicators for the second partition (truth should be put here if any)}
}
\value{
a list with adjusted Rand index (large = agreement), sensitivity and specificity, and Variation
of information (0 means perfect agreement)
}
\description{
Compare two partitions (could be of different lengths)
}

