% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pair2id}
\alias{pair2id}
\title{This function convert ids: (will need to move to the spotgear package.)}
\usage{
pair2id(N_g_vec, set_id, lane_id, include_orig_lane1 = TRUE)
}
\arguments{
\item{N_g_vec}{A vector of integers representing the number of lanes on each gel}

\item{set_id}{a vector of set ids}

\item{lane_id}{a vector of lane ids (has the same length as set_id)}

\item{include_orig_lane1}{Default to \code{TRUE} - include lane 1 while
converting; Otherwise set to \code{FALSE}.}
}
\value{
A vector of ids upon stacking lanes on one another
}
\description{
This function convert ids: (will need to move to the spotgear package.)
}
\examples{
N_g_vec     <- c(20,18,20,20,20,20,   20,20,20,9,   20,20,20) 
# <-- number of lanes per gel.
index_start <- c(1,7,11) # the 1st, 7th, 11th gel that represents 
 #the first of each cancer group.
set_no_vec  <- c(1:6,1:4,1:3) 
# index_start <- c(1,7,11,14)
# set_no_vec <- c(1:6,1:4,1:3,1:4)
PMSCL_machine_lane_id <- c(20,18,16,20,2,20,   20,20,20,NA,   
2,20,2,NA,NA,NA,NA) # NB: 2nd one has deleted two empty lanes.
# ?? no ref_lanes? Note that we call the first lane to be the 
# reference lanes, sometimes we call them lane 1s.
# controls do not have identical bands: (lesson - clear documentation 
# of the variable names after the first test of the model.)
# need two ids to identify the control lanes (set id, lane id - prior 
# to removal of the first lane).
control_set_id  <- c(6,6,6,6,6,6,6,6,6,6, 
                     10,10,10,10,
                     13,13,13,13
)
control_lane_id <- c(10,11,12,13,14,15,16,17,18,19,
                     3,4,5,6,
                     17,18,19,20
)

# the control id if all the lanes are stacked together:
ctrl_in_id <- pair2id(N_g_vec,control_set_id,control_lane_id) 
#ctrl_in_id <- c(102:111+6,171:174+10,231:234+13) 
# <-- including the original lane 1s.

}
