% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_lane}
\alias{plot_lane}
\title{Plot intensity of lanes (for raw data prior to cropping)}
\usage{
plot_lane(dat, mark, lanes, threshold = 0, add_mark = TRUE, ...)
}
\arguments{
\item{dat}{The data for plotting}

\item{mark}{The handpicked bands}

\item{lanes}{The index of lanes to be plotted}

\item{threshold}{Default is 0. Plot data above this threshold value.}

\item{add_mark}{Default is \code{TRUE}. Add handpicked marks into the plot.}

\item{...}{Other graphical parameters.}
}
\value{
plot with the lanes stacked together that could be marked with
hand-picked bands
}
\description{
Can be used for binned data
}
