% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewarp.R
\name{pwl}
\alias{pwl}
\title{Piecewise linear warping function for one index in the reference}
\usage{
pwl(i, knots_ref, knots_query)
}
\arguments{
\item{i}{The i-th bin in the reference}

\item{knots_ref}{The knots for the reference}

\item{knots_query}{The knots for the lane that is to be aligned to the reference}
}
\value{
The index in the query lane that is to be aligned to the i-th bin in the reference lane
}
\description{
Do piecewise stretch and compression by matching knots_query to knots_ref, using
the latter as anchors. The entire stretch and compression is done by \code{\link{correct_batch_effect_by_ref}}.
}
