% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewarp.R
\name{pwl_after_dewarping}
\alias{pwl_after_dewarping}
\title{Piecewise linear warping after aligning the landmarks}
\usage{
pwl_after_dewarping(g_id, l_id, normalized_rf, dat_without_lane1, data_lcm_all,
  zero_to_one_v_landmark, bugs_gel_id, bugs_in_gel_id, N_per_gel, Z_map, n_vec)
}
\arguments{
\item{g_id}{Gel id}

\item{l_id}{Lane id}

\item{normalized_rf}{The normalized_rf that is between 0 and 1. Note this subtle change compared
after cropping the right end of the gel.}

\item{dat_without_lane1}{The data frame without peaks obtained from the reference lanes. Because
when doing warping, we know these reference molecules do not exactly correspond to the
molecules that would be observed in the serum samples.}

\item{data_lcm_all}{A binary matrix of the number of rows being the number of serum samples
and the number of columns being J+2 (J for intermediate landmarks and 2 end points)}

\item{zero_to_one_v_landmark}{The landmark location on the zero to one scale.}

\item{bugs_gel_id}{Starts from 1.}

\item{bugs_in_gel_id}{Starts from 1 to say 19.}

\item{N_per_gel}{Analyzed number of lanes per gel.}

\item{Z_map}{a vector of length equal to total number of peaks detected; each element 
represents the landmark a peak is aligned to}

\item{n_vec}{A vector of length equal to the total lane number; each element is the number of peaks within a lane.}
}
\value{
A list of two elements:
\itemize{
\item \code{ref_index} Among \code{normalized_rf}, the index of the one that is
closest to each of the peak-matched landmarks. Because we want to use these index as anchors to do
piecewise linear warping.
\item \code{warped_ind} Warping mapping: bin index for post-dewarping data |-> bin index for the pre-dewarping data.
}
}
\description{
Because at the right side of the gel, we have few peaks, data contain few
information about the dewarping function. We downsample the gel, so that we only
need to consider the dewarping results on the J=50 landmarks' dewarping results.
NB: assumed that all gels have 19 non-reference lanes. Need to change otherwise.
}

