% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-peak.R
\name{sign_score}
\alias{sign_score}
\title{Calculate relative scores for peak calling}
\usage{
sign_score(y, half_width, elevation)
}
\arguments{
\item{y}{The binned intensity data}

\item{half_width}{The half window size used in \code{\link{edges}}}

\item{elevation}{The minimum intensity difference a peak needs to exceed the minimum
intensity within the window defined by \code{\link{edges}}}
}
\value{
A vector of scores, one for each intensity value in a bin
}
\description{
Sign calculation for each bin in the binned intensity data. It is used
to define contiguous regions in which we will find a peak.
}

