% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{smooth_gel}
\alias{smooth_gel}
\title{Initial smoothing using the obtained raw intensities for all sample lanes in one gel}
\usage{
smooth_gel(dat_raw, span)
}
\arguments{
\item{dat_raw}{Raw data obtained from \code{\link{read_IPdata}}. First column represents
a equal-spaced grid over 0 and 1; Other columns record intensity values obtained from all 
the tested sample lanes.}

\item{span}{The parameter tha controls the degree of smoothing. See \code{\link[stats]{loess}}}
}
\value{
A list comprising two elements, \code{res} for smoothed intensities, one column
per sample lane, and \code{sigma} for the standard deviation
}
\description{
Initial smoothing using the obtained raw intensities for all sample lanes in one gel
}

